refactor: remove createTokenProvider helper function
Remove the createTokenProvider helper and let users implement their own
caching strategies. Rename AuthProvider to TokenProvider for clarity.

- Remove createTokenProvider function and export
- Rename AuthProvider type to TokenProvider
- Update tests to focus on TokenProvider interface
- Show caching example in dynamic-auth.ts instead of built-in helper
- Update client.ts to use TokenProvider type
