+++ packages/mcp/src/index.ts
@@ -1,3 +1,3 @@
 export * from './client';
 export * from './server';
-export { createTokenProvider, type AuthProvider } from './auth/token-provider';
+export { type TokenProvider } from './auth/token-provider';
